implement separate Markdown component
Content component is split up to Markdown component and actual Content.
Changed the meaning of Content component to be an actual content to improve
readability of text.

Property 'markdown' is deprecated and kept temporary for backward
compatibility. All the usages of Content must be replaced to Markdown component,
bare or wrapped into Content where it's necessary.
