+++ components/index.js
@@ -13,3 +13,4 @@ export { default as Switcher, SwitcherItem } from './switcher'
 export { default as Collapsed } from './collapsed'
 export * from './content'
 export * from './elements'
+export { default as Markdown } from './markdown'
